Moved layers to it's own folder
