+++ Eclipse/Eclipse.vcxproj
@@ -168,8 +168,8 @@
     <ClInclude Include="src\eclipse\events\mouse_events.h" />
     <ClInclude Include="src\eclipse\events\window_events.h" />
     <ClInclude Include="src\eclipse\imgui\imgui_layer.h" />
-    <ClInclude Include="src\eclipse\layer.h" />
-    <ClInclude Include="src\eclipse\layer_stack.h" />
+    <ClInclude Include="src\eclipse\layer\layer.h" />
+    <ClInclude Include="src\eclipse\layer\layer_stack.h" />
     <ClInclude Include="src\eclipse\log.h" />
     <ClInclude Include="src\eclipse\window.h" />
     <ClInclude Include="src\ecpch.h" />
@@ -179,8 +179,8 @@
   <ItemGroup>
     <ClCompile Include="src\eclipse\application.cpp" />
     <ClCompile Include="src\eclipse\imgui\imgui_layer.cpp" />
-    <ClCompile Include="src\eclipse\layer.cpp" />
-    <ClCompile Include="src\eclipse\layer_stack.cpp" />
+    <ClCompile Include="src\eclipse\layer\layer.cpp" />
+    <ClCompile Include="src\eclipse\layer\layer_stack.cpp" />
     <ClCompile Include="src\eclipse\log.cpp" />
     <ClCompile Include="src\ecpch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
