+++ Eclipse/src/Eclipse.h
@@ -3,7 +3,7 @@
 // For use by Eclipse applications
 
 #include "eclipse/application.h"
-#include "eclipse/layer.h"
+#include "eclipse/layer/layer.h"
 #include "eclipse/log.h"
 #include "Eclipse/imgui/imgui_layer.h"
