+++ Eclipse/src/eclipse/imgui/imgui_layer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "eclipse/layer.h"
+#include "eclipse/layer/layer.h"
 
 namespace eclipse {
