+++ Eclipse/src/eclipse/layer/layer_stack.cpp
