+++ Eclipse/src/eclipse/layer/layer_stack.h
